+++ test-integration/00-install-all-jhipster.sh
@@ -23,4 +23,4 @@ npm install -g yo
 
 echo "*** generator-jhipster-nodejs: use current branch version"
 npm install -g
-# npm link
+# sudo npm link