+++ test/nestjs-service.spec.js
@@ -3,6 +3,10 @@ const fse = require('fs-extra');
 const assert = require('yeoman-assert');
 const helpers = require('yeoman-test');
 
+const constants = require('../generators/generator-nodejs-constants');
+
+const SERVER_NODEJS_DIR = `${constants.SERVER_NODEJS_SRC_DIR}/`;
+
 describe('Subgenerator nestjs-service of nodejs JHipster blueprint', () => {
     describe('1-test', () => {
         before(done => {
@@ -10,6 +14,7 @@ describe('Subgenerator nestjs-service of nodejs JHipster blueprint', () => {
                 .run('generator-jhipster/generators/spring-service')
                 .inTmpDir(dir => {
                     fse.copySync(path.join(__dirname, '../test/templates/ngx-blueprint'), dir);
+                    fse.copySync(path.join(__dirname, '../test/templates/server'), `${dir}/server`);
                 })
                 .withOptions({
                     'from-cli': true,
@@ -24,16 +29,20 @@ describe('Subgenerator nestjs-service of nodejs JHipster blueprint', () => {
                         path.join(__dirname, '../generators/spring-service/index.js')
                     ]
                 ])
-                .withArguments(['foo'])
+                .withArguments(['bar'])
                 .withPrompts({
                     useInterface: false
                 })
                 .on('end', done);
         });
 
-        it('it works', () => {
+        it('app exists with bar service ts file', () => {
             // Adds your tests here
-            assert.textEqual('Write your own tests!', 'Write your own tests!');
+            assert.file(`${SERVER_NODEJS_DIR}src/service/bar.service.ts`);
+
+            const testServicePath = `${SERVER_NODEJS_DIR}test/service/bar.service.spec.ts`;
+            assert.file(testServicePath);
+            assert.fileContent(testServicePath, 'BarService');
         });
     });
 });