+++ test/templates/server/src/app.module.ts
@@ -1,27 +1,20 @@
-import { Module, NestModule } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { JwtStrategy } from './security/passport.jwt.strategy';
-import { UsersModule } from './module/users.module';
+import { Module } from '@nestjs/common';
+import { AuthModule } from './module/auth.module';
 // jhipster-needle-add-entity-module-to-main-import - JHipster will import entity modules here, do not remove
 // jhipster-needle-add-controller-module-to-main-import - JHipster will import controller modules here, do not remove
+// jhipster-needle-add-service-module-to-main-import - JHipster will import service modules here, do not remove
 
 @Module({
   imports: [
-    TypeOrmModule.forRoot({
-      name: 'default',
-      type: 'sqlite',
-      database: '../target/sqlite-dev-db.sql',
-      synchronize: true,
-      logging: false,
-      entities: [__dirname + '/domain/*.entity{.ts,.js}']
-    }),
-    UsersModule
+    AuthModule,
   // jhipster-needle-add-entity-module-to-main - JHipster will add entity modules here, do not remove
   ],
   controllers: [
   // jhipster-needle-add-controller-module-to-main - JHipster will add controller modules here, do not remove
   ],
-  providers: [JwtStrategy]
+  providers: [
+    // jhipster-needle-add-service-module-to-main - JHipster will add service modules here, do not remove
+  ]
 })
 export class AppModule  {