+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+<a name="1.0.0"></a>
 <a name="1.0.0-beta.2"></a>
 <a name="1.0.0-beta.1"></a>
 <a name="1.0.0-alpha.3"></a>
 <a name="1.0.0-alpha.2"></a>
+
+# [1.0.0](https://github.com/jhipster/generator-jhipster-nodejs/tree/v1.0.0)
+
+-   Provide service generator [issue #94](https://github.com/jhipster/generator-jhipster-nodejs/issues/94)
+-   Check controller generator with lint and tests [issue #92](https://github.com/jhipster/generator-jhipster-nodejs/issues/92)
 
 # [1.0.0-beta.2](https://github.com/jhipster/generator-jhipster-nodejs/tree/v1.0.0-beta.2)